Move first-run setup out of core
